requirements.txt
openai
requests
dotenv
streamlit
scikit-learn
matplotlib
plotly
pyyaml
mlflow